refactored generate string
