+++ express_server.js
@@ -8,6 +8,7 @@ const PORT = 8080;
 
 // HELPERS
 const lookupIDByEmail = require('./helpers');
+const generateRandomString = require('./helpers');
 
 app.set('view engine', 'ejs');
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -179,14 +180,3 @@ app.listen(PORT, () => {
   console.log(`Server Listening on Port ${PORT}`);
 });
 
-const generateRandomString = function () {
-  const charset = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
-  const stringLength = 6;
-  let shortURL = '';
-  let $n = 0;
-  for (let i = 0; i < stringLength; i++) {
-    $n = Math.floor(Math.random() * charset.length);
-    shortURL += charset[$n];
-  }
-  return shortURL;
-};
