+++ helpers.js
@@ -7,4 +7,17 @@ const lookupIDByEmail = function (email, userDB) {
   return false;
 };
 
+const generateRandomString = function () {
+  const charset = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
+  const stringLength = 6;
+  let shortURL = '';
+  let $n = 0;
+  for (let i = 0; i < stringLength; i++) {
+    $n = Math.floor(Math.random() * charset.length);
+    shortURL += charset[$n];
+  }
+  return shortURL;
+};
+
 module.exports = lookupIDByEmail;
+module.exports = generateRandomString;
